requirements.txt
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
biopython==1.68
certifi==2019.11.28
Jinja2==2.8
MarkupSafe==0.23
numpy==1.12.1
pandas==0.18.1
python-dateutil==2.5.3
pytz==2016.4
scikit-learn==0.18
scipy==0.19.1
singledispatch==3.4.0.3
six==1.10.0
tornado==4.3
